Start to deal with the building on the grid and instantiated them from a cvs file
In this commit a big change has been made. Basically we build a finite
list of 24 building in buildings and then we have a list called
buildinsOnGrid that will be updated according to what is sent by the
scanner. So we make a clear difference between the physical existing 24
building (that may evolve) and the one present on the grid.